honda_restaurant/requirements.txt
Django==3.1.7
djangorestframework==3.12.2
drf-yasg==1.20.0
psycopg2-binary==2.8.6